requirements.txt
streamlit
pandas
finance-datareader
beautifulsoup4
matplotlib 
plotly
openpyxl
html5lib
pymysql
openai
python-dotenv